Remove init functions (#507)
* feat: moved commands from being initialized from init functions to classic create functions

* Align command creation names
