+++ pkg/rest/rest_caller.go
@@ -41,6 +41,7 @@ type RestCallerSettings interface {
 	AppId() string
 	RestAdaptedAppId() string
 	PrintMode() log.PrintMode
+	IsSenseForKubernetes() bool
 }
 
 type RestCaller struct {
